+++ modules/bench/src/bench.js
@@ -243,7 +243,13 @@ async function runBenchTestAsync(test) {
   let totalIterations = 0;
 
   for (let i = 0; i < test.opts.minIterations; i++) {
-    const {time, iterations} = await runBenchTestTimedAsync(test, test.opts.time);
+    let time;
+    let iterations;
+    if (test.async && test.throughput) {
+      ({time, iterations} = await runBenchTestParallelIterationsAsync(test, test.throughput));
+    } else {
+      ({time, iterations} = await runBenchTestForMinimumTimeAsync(test, test.opts.time));
+    }
     const iterationsPerSecond = iterations / time;
     results.push(iterationsPerSecond);
     totalTime += time;
@@ -259,7 +265,7 @@ async function runBenchTestAsync(test) {
 }
 
 // Run a test func for an increasing amount of iterations until time threshold exceeded
-async function runBenchTestTimedAsync(test, minTime) {
+async function runBenchTestForMinimumTimeAsync(test, minTime) {
   let iterations = 1;
   let elapsedMillis = 0;
 
@@ -288,19 +294,26 @@ async function runBenchTestTimedAsync(test, minTime) {
   };
 }
 
+// Run a test func for a specific amount of parallel iterations
+async function runBenchTestParallelIterationsAsync(test, iterations) {
+  const testArgs = test.initFunc && test.initFunc();
+
+  const promises = [];
+
+  const {context, testFunc} = test;
+  for (let i = 0; i < iterations; i++) {
+    promises.push(testFunc.call(context, testArgs));
+  }
+
+  return await Promise.all(promises);
+}
+
 // Run a test func for a specific amount of iterations
 async function runBenchTestIterationsAsync(test, iterations) {
   const testArgs = test.initFunc && test.initFunc();
-
   const {context, testFunc} = test;
-  if (context && testArgs) {
-    for (let i = 0; i < iterations; i++) {
-      await testFunc.call(context, testArgs);
-    }
-  } else {
-    for (let i = 0; i < iterations; i++) {
-      await testFunc.call(context);
-    }
+  for (let i = 0; i < iterations; i++) {
+    await testFunc.call(context, testArgs);
   }
 }
 
@@ -310,6 +323,7 @@ async function runBenchTestIterationsAsync(test, iterations) {
 function runBenchTestIterations(test, iterations) {
   const testArgs = test.initFunc && test.initFunc();
 
+  // When running sync, avoid overhead of parameter passing if not needed
   const {context, testFunc} = test;
   if (context && testArgs) {
     for (let i = 0; i < iterations; i++) {